chore : Difficulty move dir
