+++ src/main/java/com/example/project/solution/dto/request/admin/update/SolutionUpdateRequest.java
@@ -1,6 +1,6 @@
 package com.example.project.solution.dto.request.admin.update;
 
-import com.example.project.solution.domain.Difficulty;
+import com.example.project.solution.domain.vo.Difficulty;
 import lombok.Data;
 
 @Data
